[~] Folder restructuration of the core and edit layer_list() return string format
