+++ src/Hylozoa-Engine/Components/Rendering/Serializer.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "Hylozoa-Engine/Core/LayerManager.hpp"
+#include "../../Core/Layers/LayerManager.hpp"
 #include "Renderable.hpp"
 
 #include <nlohmann/json.hpp>
